requirements.txt
fastapi==0.96.0
pydantic==1.10.8
SQLAlchemy==2.0.15
uvicorn==0.22.0
bcrypt==4.0.1
mysql-connector-python~=8.0.33
python-dotenv==1.0.0
PyJWT~=2.7.0